pmOS: backlight: qcom-wled: Use sink_addr for sync toggle
https://gitlab.com/sdm845-mainline/sdm845-linux/-/commit/803e65cc3c7d5b12adbcb47c00232aa9cfc1c5f7

WLED3_SINK_REG_SYNC is, according to the name, a sink register offset.
Therefore, use the sink address as base.

This fixes the sync toggle on wled4, which can be
observed by the fact that adjusting brightness works.

It has no effect on wled3 because sink and ctrl base addresses are the same.
This allows adjusting the brightness without
having to disable then enable the module.

Signed-off-by: Obeida Shamoun <oshmoun100@googlemail.com>
Signed-off-by: Konrad Dybcio <konradybcio@gmail.com>
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>
Change-Id: Ieb4233e90fd7b5881bef3e981ae6edac11874ef5